Split Settings collection resources
